+++ speed-daemon/src/handlers/error.rs
@@ -1,0 +1,16 @@
+
+use tokio;
+use speed_daemon::message::OutboundMessageType;
+use tokio::sync::mpsc;
+
+pub fn handle_error(
+    error_message: String,
+    tx: mpsc::Sender<OutboundMessageType>,
+) -> anyhow::Result<()> {
+    tokio::spawn(async move {
+        tx.send(OutboundMessageType::Error(error_message))
+            .await
+            .expect("Unable to send error message");
+    });
+    Ok(())
+}
