+++ speed-daemon/src/handlers/mod.rs
@@ -1,7 +1,11 @@
 mod dispatcher;
 mod heartbeat;
 mod camera;
+mod plate;
+mod error;
 
 pub use dispatcher::handle_i_am_dispatcher;
 pub use heartbeat::handle_want_hearbeat;
 pub use camera::handle_i_am_camera;
+pub use plate::handle_plate;
+pub use error::handle_error;
